[Feat] 소모아이템 사용 시 수치 적용 및 사용 아이템 제거, UI 디자인 마무리

+++ Assets/Scripts/UI/PlayerStatUI.cs
@@ -63,6 +63,6 @@ public class PlayerStatUI : MonoBehaviour
 
     public void SetExpBar(int curExp, int levelUpValue)
     {
-        expBar.fillAmount = curExp / levelUpValue;
+        expBar.fillAmount = curExp / (float)levelUpValue;
     }
 }
